pyproject.toml
[build-system]
requires = [
    "scikit-build-core",
    "numpy",
    "meson",
    "charset_normalizer"
]
build-backend = "scikit_build_core.build"


[project]
dependencies = [
    "astropy>=6.1.0",
    "numpy>=1.26.0",
    "pyyaml>=6.0.2",
    "scipy>=1.12.0",
    "requests>=2.25",
    "spiceypy>=6.0.0",
    "unlzw3>=0.2.2",
    "ply>=3.11",
    "multidict>=6.1.0",
    "matplotlib>=3.9.0"
]
name = "pride"
dynamic = [
    "version"
]
description = ""
authors = [
    { name = "Alfonso Sánchez Rodríguez", email = "a.sanchezrodriguez@student.tudelft.nl" }
]
requires-python = ">=3.11"

[tool.scikit-build]
metadata.version.provider = "scikit_build_core.metadata.setuptools_scm"
sdist.include = [
    "src/pride/_version.py"
]
ninja.version = ">=1.10"
cmake.version = ">=3.17.2"

[tool.setuptools_scm] # Section required
write_to = "src/pride/_version.py"
